requirements.txt
crispy-bootstrap5==2024.10
Django==5.1.2
django-braces==1.16.0
django-ckeditor==6.7.1
django-crispy-forms==2.3
pillow==10.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.9

